requirements.txt
kubernetes==28.1.0
fastapi~=0.103.1
uvicorn==0.23.2
starlette~=0.27.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.21
psycopg2-binary==2.9.9
PyYAML~=6.0.1